Move get_countrytag to DO
